Moved EnabledChecker into FancyHologramsPlugin
